refactor: split unit-order dialogs and naval panel list parts (#3878)
Extract move-fleet pick builders, move-army declare-war confirm flow,
and naval panel tree list projection into library parts; document
paths in repo-and-packages and extend the event-bus allow-list.
