+++ tool/check_app_event_bus_decoupling.dart
@@ -123,6 +123,7 @@ const Set<String> _allowedFeatureLocalDialogFiles = <String>{
   // (split / move army; invasion confirm sub-dialog of move army).
   'app/lib/features/game/widgets/units/military/military_units_panel.dart',
   'app/lib/features/game/widgets/unit_orders/move_army_dialog.dart',
+  'app/lib/features/game/widgets/unit_orders/move_army_dialog_declare_war.dart',
   // Deferred per #2626 scope (game-side menu game-parameters dialog and
   // production breakdown). Migrating these to typed bus events is
   // explicitly out of scope for #2626 and must be filed as separate
